122. 1. Let's set up the Core Data project with SwiftUI and fix Xcode's template
